requirements.txt
Django==4.1.3
djangorestframework==3.14.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
Pillow==9.3.0
asgiref==3.5.2
django-filter-22.1
djoser==2.1.0
